Pulling out class in to own file.
