+++ src/CompaniesHouse/CompaniesHouseCompanyInsolvencyInformationClient.cs
@@ -2,7 +2,6 @@ using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
 using CompaniesHouse.Response.Insolvency;
-using Newtonsoft.Json;
 
 namespace CompaniesHouse
 {
@@ -26,19 +25,6 @@ namespace CompaniesHouse
             var result = await response.Content.ReadAsJsonAsync<CompanyInsolvencyInformation>().ConfigureAwait(false);
 
             return new CompaniesHouseClientResponse<CompanyInsolvencyInformation>(result);
-        }
-    }
-
-    public static class HttpContentExtensions
-    {
-        public static async Task<T> ReadAsJsonAsync<T>(this HttpContent content)
-        {
-            var input = await content.ReadAsStringAsync()
-                .ConfigureAwait(false);
-
-            var obj = JsonConvert.DeserializeObject<T>(input);
-
-            return obj;
         }
     }
 }
